requirements.txt
spacy-llm=0.1.0
python_dotenv=1.0.0
jupyterlab=3.6.3
pygls=1.0.1